Clase 7 - Forma correcta de realizar un RESPONSER

+++ apps/users/api/api.py
@@ -1,3 +1,4 @@
+from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
@@ -18,7 +19,7 @@ def user_api_view(request):
         #Se hace el tratamiento de la información, se pone many=True porque va a traer muchos archivos
         user_serializer = UserSerializer(users, many= True)
         #Se retorna con response a la vista, PERO la informacion se encuentra en el atributo .data
-        return Response(user_serializer.data)
+        return Response(user_serializer.data, status = status.HTTP_200_OK)
     
     #Cuando se hace un POST, se puede usar el serializer para corroborar que la info sea correcta
     elif request.method == 'POST':
@@ -28,29 +29,33 @@ def user_api_view(request):
         if user_serializer.is_valid():
             #Se guarda la informacion en la base de datos
             user_serializer.save()
-            return Response(user_serializer.data)
-        return Response(user_serializer.errors)
+            return Response({'message':'Usuario creado correctamente!'}, status = status.HTTP_201_CREATED)
+        return Response(user_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
 
 #Se crea la vista de detalle de cada usuario, tambien se le agrega los metodos PUT y DELETE
 @api_view(['GET', 'PUT', 'DELETE'])
 def user_detail_api_view(request, pk=None):
+    #Queryset, a la funcion se le ingresa el pk, o el parametro de consulta, que se va desde la url
+    user = User.objects.filter(id = pk).first()
+    #Se valida si el usuario existe o no
+    if user:
 
-    #A la funcion se le ingresa el pk, o el parametro de consulta, que se va desde la url
-    #Se define cada uno de los metodos, en GET se muestra el detalle
-    if request.method == 'GET':
-        user = User.objects.filter(id = pk).first()
-        user_serializer = UserSerializer(user)
-        return Response(user_serializer.data)
-    #Este se utiliza para acutalizar la info, comprueba por medio del serializer la info enviada
-    elif request.method == 'PUT':
-        user = User.objects.filter(id = pk).first()
-        user_serializer = UserSerializer(user, data = request.data)
-        if user_serializer.is_valid():
-            user_serializer.save()
-            return Response(user_serializer.data)
-        return Response(user_serializer.errors)
-    #Elimina la información
-    elif request.method == 'DELETE':
-        user = User.objects.filter(id = pk).first()
-        user.delete()
-        return Response('Eliminado')
+        #Se define cada uno de los metodos, en GET se muestra el detalle
+        if request.method == 'GET':
+            user_serializer = UserSerializer(user)
+            return Response(user_serializer.data, status = status.HTTP_200_OK)
+        
+        #Este se utiliza para acutalizar la info, comprueba por medio del serializer la info enviada
+        elif request.method == 'PUT':
+            user_serializer = UserSerializer(user, data = request.data)
+            if user_serializer.is_valid():
+                user_serializer.save()
+                return Response(user_serializer.data, status = status.HTTP_200_OK)
+            return Response(user_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
+        
+        #Elimina la información
+        elif request.method == 'DELETE':
+            user.delete()
+            return Response({'message':'Usuario eliminado correctamente!'}, status = status.HTTP_200_OK)
+    #Retorna un mensaje de error
+    return Response({'message':'No se ha encontrado un usuario con estos datos'}, status = status.HTTP_400_BAD_REQUEST)
